+++ packages/app_package_publisher_firebase/lib/src/publish_firebase_config.dart
@@ -69,7 +69,6 @@ class PublishFirebaseConfig extends PublishConfig {
     cmdData.forEach((key, value) {
       cmdList.addAll([key, value!]);
     });
-    // print("cmdList:${cmdList.toString()}");
     return cmdList;
   }
 }