+++ packages/flutter_app_publisher/CHANGELOG.md
@@ -1,6 +1,11 @@
+## 0.1.5
+
+- Add `firebase` publisher.
+- Add `github` publisher.
+
 ## 0.1.4
 
-- Add qiniu publisher.
+- Add `qiniu` publisher.
 
 ## 0.1.0